Add support for Universal binaries on OS X
Most dynamic libraries on OS X are in the form of Universal
binaries. These are also called "Fat binaries" and their
specification can be found here:
 https://developer.apple.com/library/mac/documentation/developertools/conceptual/MachORuntime/Reference/reference.html#jumpTo_39

This adds support for loading these files and locating the
binaries bundled inside the fat binary. It only support Mach-O
binaries right now as this is most common, but more formats could
easily be added.

Some methods such as getEntryPoint will only return the entry point
for the first binary. This is because ROPgadget currently expects
to work on only a single binary. If support for working with
multiple binaries at the same time will ever be added, it should be
fairly easy to return a list of entry points instead.