Filter out useless RISC-V gadgets
This filters out gadgets like "andi sp, sp, -0x10", which do not
perform a jump, but whose last two bytes ("01ff" in this case) are
a compressed jump instruction ("c.bnez a4, -0xe8" in this case).

We perform this filtering only on RISC-V because it leads to missed
gadgets in x86, where prefixes can be added to an instruction that change
its size but not its behavior. MIPS is especially problematic because we
set gad_size = 8 in order to display the next instruction in the branch
delay slot. On other architectures this filter makes no difference on
the test suite.